Use SetHandler instead of ProxyPassMatch to pass php files to fcgi process
Use SetHandler instead of the ProxyPassMatch configuration as the ProxyPassMatch
takes precedence over stuff like mod_rewrite as it is handled first. That means
rewrite rules written for *.php files do not have any effect as the request
is sent to the "proxy" and not handled by the other directives like mod_rewrite.